requirements.txt
asgiref==3.5.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.12
django-cors-headers==3.11.0
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
ec2-hibinit-agent==1.0.0
entrypoints==0.3
gunicorn==20.1.0
hibagent==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
itypes==1.2.0
Jinja2==3.1.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.4.0
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.3
python-debian===0.1.36ubuntu1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.17.6
six==1.14.0
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
typing-extensions==4.1.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
uWSGI==2.0.20
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
